+++ haskell/lib/Error.hs
@@ -15,3 +15,5 @@ assignError = MkUnknownError
 -- TODO: add megaparsec to parse Error Messages
 --
 -- TODO: add error when internet connection is off
+--
+-- MkSystemError 1 "/home/xecarlox/Desktop/p2p-rendering-computation/p2p-rendering-computation" "2024/11/09 21:08:06 Get \"http://0.0.0.0:8088/MAPPort?port=3333&domain_name=\": dial tcp 0.0.0.0:8088: connect: connection refused\n"
