+++ haskell/src/Main.hs
@@ -1,9 +1,10 @@
 
+
 module Main where
 
 import P2Prc (runP2Prc)
 
 
 main :: IO ()
-main = do
+main =
   runP2Prc
